refactored startProcess function
